Changes UInt64 to size_t to avoid having to specify size arguments

+++ guttural/include/gut_array.h
@@ -8,8 +8,8 @@
 typedef struct GutturalArray
 {
     GutTValue * elements;
-    UInt64 capacity;
-    UInt64 count;
+    size_t capacity;
+    size_t count;
 } GutArray;
